Restructured everything to work with Blueprints
