--- .gitignore
@@ -1,5 +1,0 @@
-# En tu archivo .gitignore
-.vscode/*
-*.pyc
-*.pyo
-*.pyd
