--- Module/Login/loginController/loginLogic.py
@@ -1,6 +1,0 @@
-from Model.loginModel.userCredentials import User
-
-testUser = User("Raul", "Sekiro")
-
-def authenticateCredentials(username, password):
-    return testUser.validate(username, password)
