+++ Module/Login/loginViews/login.html
@@ -20,7 +20,7 @@
 
       <h2>Corporation Credentials</h2>
 
-      <form action="{{ url_for('login') }}" method="post" class="form">
+      <form action="{{ url_for('login_bp.login') }}" method="post" class="form">
 
         <div class="form__field">
           <input type="text" placeholder="Usuario" name="username">
